Add custom molecule file to cookiecutter
One of the assumptions of testing is that the cookiecutter template
can be executed without changes.

This also feels like a good approach as it provides documentation of
usage and helps guide first users.

The standard molecule file has an empty platforms section
This doesn't document usage, it also causes create.yml to throw errors

We provide a default platform with a recent Ubuntu image
Ubuntu as it mimics the defaults in create and the documentation
Image owner & name syntax allows us to be region agnostic