Fix openssh_keypair Ansible 2.9 compatibility
crypto.openssh_keypair regenerate option added post-2.9

Removing it could cause slight complications with specifying an existing key.

regenerate has three options we care about:
* never - current setting. Existing key will never be replaced
* partial_idempotence - 3.0 setting. Key is replaced if options don't match
* full_idempotence - 2.9 setting. partial, and if key has unknown passphrase

Never would be preferable, but 2.9 support is required.
I can not see a method of adding the parameter only for 3.0

To reduce the risks, the type and size parameters were also removed.
The default type is the same, the size is bigger.
The significant impact however is that if there is an existing key with
different values, that key will not be replaced.

The documentation was also updated to reflect the risk with the
passphrase.